requirements.txt
streamlit
langchain
langchain-groq
langchain-core
httpx
python-dotenv
pydantic